requirements.txt
gradio==4.44.0
openai==1.51.0
pillow==10.2.0
pyyaml==6.0.1